Fix build
Convert libc-test and musl into git submodules so we can update as needed
without huge changesets.

Using submodules means we can just deleted the .c files we convert, so we
provide a file called ported_objects which contains a list of .o files to
delete from libc.a before adding in librusl.a's contents.

We include some discussion of symbol conflicts and resolution in the readme.

We also sneak in a couple rusl changes:
    * privatize a few malloc.rs symbols because they aren't public in musl
    * correct clock_setttime to clock_settime