Improve scripts
The major change here is some script refactoring and the addition of the full
libc-test run using musl's libc.a before we mix in librusl.a's symbols. That
is, we run libc-test _twice_: once using musl's original symbols, and again
using rusl's symbols.

We run the tests using musl so we have an up-to-date record of what fails on
a given machine, rather than using baseline_failures, which must be kept up-
to-date and runs the risk of being wrong on different machines.